# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vartalaap API",
    "description": "Voice bot platform for local Indian businesses",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/call-logs": {
      "get": {
        "tags": [
          "Call Logs",
          "Call Logs"
        ],
        "summary": "List Call Logs",
        "description": "List call logs with optional filters.\n\nSecurity: business_id is required to prevent cross-tenant data access.",
        "operationId": "list_call_logs_api_call_logs_get",
        "parameters": [
          {
            "name": "business_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Required for tenant isolation",
              "title": "Business Id"
            },
            "description": "Required for tenant isolation"
          },
          {
            "name": "outcome",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CallOutcome"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CallLogResponse"
                  },
                  "title": "Response List Call Logs Api Call Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/call-logs/summary": {
      "get": {
        "tags": [
          "Call Logs",
          "Call Logs"
        ],
        "summary": "Get Call Log Summary",
        "description": "Get summary statistics for call logs.\n\nSecurity: business_id is required to prevent cross-tenant data access.",
        "operationId": "get_call_log_summary_api_call_logs_summary_get",
        "parameters": [
          {
            "name": "business_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Required for tenant isolation",
              "title": "Business Id"
            },
            "description": "Required for tenant isolation"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallLogSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/call-logs/{call_log_id}": {
      "get": {
        "tags": [
          "Call Logs",
          "Call Logs"
        ],
        "summary": "Get Call Log",
        "description": "Get a call log by ID.",
        "operationId": "get_call_log_api_call_logs__call_log_id__get",
        "parameters": [
          {
            "name": "call_log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Call Log Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallLogResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CallLogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "business_id": {
            "type": "string",
            "title": "Business Id"
          },
          "caller_id_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caller Id Hash"
          },
          "call_start": {
            "type": "string",
            "title": "Call Start"
          },
          "call_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Call End"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "detected_language": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DetectedLanguage"
              },
              {
                "type": "null"
              }
            ]
          },
          "transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript"
          },
          "extracted_info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Info"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CallOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "consent_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "business_id",
          "caller_id_hash",
          "call_start",
          "call_end",
          "duration_seconds",
          "detected_language",
          "transcript",
          "extracted_info",
          "outcome",
          "consent_type",
          "created_at"
        ],
        "title": "CallLogResponse",
        "description": "Response schema for call logs."
      },
      "CallLogSummary": {
        "properties": {
          "total_calls": {
            "type": "integer",
            "title": "Total Calls"
          },
          "total_duration_seconds": {
            "type": "integer",
            "title": "Total Duration Seconds"
          },
          "avg_duration_seconds": {
            "type": "number",
            "title": "Avg Duration Seconds"
          },
          "calls_by_outcome": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Calls By Outcome"
          },
          "calls_by_language": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Calls By Language"
          }
        },
        "type": "object",
        "required": [
          "total_calls",
          "total_duration_seconds",
          "avg_duration_seconds",
          "calls_by_outcome",
          "calls_by_language"
        ],
        "title": "CallLogSummary",
        "description": "Summary stats for call logs."
      },
      "CallOutcome": {
        "type": "string",
        "enum": [
          "resolved",
          "fallback",
          "dropped",
          "error",
          "privacy_opt_out"
        ],
        "title": "CallOutcome",
        "description": "How a call ended."
      },
      "ConsentType": {
        "type": "string",
        "enum": [
          "none",
          "transcript",
          "whatsapp"
        ],
        "title": "ConsentType",
        "description": "Level of consent given by caller."
      },
...
      "DetectedLanguage": {
        "type": "string",
        "enum": [
          "hindi",
          "english",
          "hinglish"
        ],
        "title": "DetectedLanguage",
        "description": "Detected language during call."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}